+++ src/Repository/Repository.php
@@ -5,6 +5,8 @@ namespace Kununu\Elasticsearch\Repository;
 
 use Elasticsearch\Client;
 use Exception;
+use InvalidArgumentException;
+use Kununu\Elasticsearch\Exception\RepositoryConfigurationException;
 use Kununu\Elasticsearch\Exception\RepositoryException;
 use Kununu\Elasticsearch\Query\Query;
 use Kununu\Elasticsearch\Query\QueryInterface;
@@ -37,7 +39,7 @@ class Repository implements RepositoryInterface, LoggerAwareInterface
     protected $config;
 
     /**
-     * ElasticsearchRepository constructor.
+     * Repository constructor.
      *
      * @param \Elasticsearch\Client $client
      * @param array                 $config
@@ -51,8 +53,25 @@ class Repository implements RepositoryInterface, LoggerAwareInterface
     /**
      * @inheritdoc
      */
-    public function save(string $id, array $document): void
+    public function save(string $id, $entity): void
     {
+        if (is_array($entity)) {
+            $document = $entity;
+        } elseif (is_object($entity)) {
+            $configuredEntityClass = $this->config->getEntityClass();
+            if ($configuredEntityClass && $entity instanceof $configuredEntityClass) {
+                $document = $entity->toElastic();
+            } elseif ($this->config->getEntitySerializer()) {
+                $document = $this->config->getEntitySerializer()->toElastic($entity);
+            } else {
+                throw new RepositoryConfigurationException(
+                    'No entity serializer configured while trying to persist object'
+                );
+            }
+        } else {
+            throw new InvalidArgumentException('Entity must be of type array or object');
+        }
+
         $this->execute(
             function () use ($id, $document) {
                 $this->client->index(
@@ -210,18 +229,6 @@ class Repository implements RepositoryInterface, LoggerAwareInterface
     }
 
     /**
-     * @param \Exception $e
-     *
-     * @throws \Kununu\Elasticsearch\Exception\RepositoryException
-     */
-    protected function logErrorAndThrowException(Exception $e): void
-    {
-        $this->getLogger()->error(self::EXCEPTION_PREFIX . $e->getMessage());
-
-        throw new RepositoryException($e->getMessage(), $e);
-    }
-
-    /**
      * @param string $operationType
      *
      * @return array
@@ -255,7 +262,31 @@ class Repository implements RepositoryInterface, LoggerAwareInterface
      */
     protected function parseRawSearchResponse(array $rawResult): ResultIteratorInterface
     {
-        return ResultIterator::create($rawResult['hits']['hits'] ?? [])
+        $results = $hits = $rawResult['hits']['hits'] ?? [];
+
+        if ($this->config->getEntityClass()) {
+            $results = array_map(
+                function (array $hit) {
+                    $metaData = $hit;
+                    unset($metaData['_source']);
+
+                    return $this->config->getEntityClass()::fromElasticDocument($hit['_source'], $metaData);
+                },
+                $hits
+            );
+        } elseif ($this->config->getEntityFactory()) {
+            $results = array_map(
+                function (array $hit) {
+                    $metaData = $hit;
+                    unset($metaData['_source']);
+
+                    return $this->config->getEntityFactory()->fromDocument($hit['_source'], $metaData);
+                },
+                $hits
+            );
+        }
+
+        return ResultIterator::create($results)
             ->setTotal($rawResult['hits']['total'] ?? 0)
             ->setScrollId($rawResult['_scroll_id'] ?? null);
     }